+++ videosnap/Constants.m
@@ -10,7 +10,7 @@
 
 NSString *const VERSION = @"0.0.8";
 
-int   const DEFAULT_FRAMES_PER_SECOND = 30;
+int const DEFAULT_FRAMES_PER_SECOND = 30;
 float const DEFAULT_RECORDING_DELAY = 0.5;
 NSString *const DEFAULT_ENCODING_PRESET = @"Medium";