+++ videosnap/main.m
@@ -15,7 +15,7 @@ BOOL isInterrupted;
 VideoSnap *videoSnap;
 
 /**
- * signal interrupt handler
+ * signal interrupt handler Ctrl+c
  */
 void SIGINT_handler(int signum) {
 	if (!isInterrupted && [videoSnap isRecording]) {
@@ -27,14 +27,25 @@ void SIGINT_handler(int signum) {
 }
 
 /**
+ * signal interrupt handler Ctrl+z
+ */
+void SIGSTP_handler(int signum) {
+    if ([videoSnap isRecording]) {
+        [videoSnap togglePauseRecording:signum];
+    }
+}
+
+/**
  * main
  */
 int main(int argc, const char * argv[]) {
-
 	isInterrupted = NO;
 	
-	// setup int handler for Ctrl+C cancelling
+	// setup int handlers
+    // Ctrl+C cancels
+    // Ctrl+Z pause/resume
 	signal(SIGINT, &SIGINT_handler);
+    signal(SIGTSTP, &SIGSTP_handler);
 
 	// C args as NSArray
 	NSMutableArray *args = [[NSMutableArray alloc] initWithCapacity: argc];
@@ -42,7 +53,8 @@ int main(int argc, const char * argv[]) {
 		[args addObject: [NSString stringWithCString: argv[i] encoding: NSUTF8StringEncoding]];
 	}
     
-    videoSnap = [[VideoSnap alloc] initWithVerbosity:([args indexOfObject:@"-v"] != NSNotFound)];
+    videoSnap = [[VideoSnap alloc]
+                 initWithVerbosity:([args indexOfObject:@"-v"] != NSNotFound)];
 
     return [videoSnap processArgs: args];
 }